makefile: ensure cross-compilation compatibility
We need to ensure the snapshotter can be cross-compiled across all
platforms, and although we currently only support linux. We must
make sure that external projects (such as buildkit) are not affected
by cross-compiling when they import package such as `converter`.
