+++ pkg/converter/convert_unix.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 /*
  * Copyright (c) 2022. Nydus Developers. All rights reserved.
  *
@@ -20,7 +23,6 @@ import (
 	"github.com/containerd/containerd/archive/compression"
 	"github.com/containerd/containerd/content"
 	"github.com/containerd/fifo"
-	"github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
 	"golang.org/x/sync/errgroup"
 
@@ -31,32 +33,6 @@ const bootstrapNameInTar = "image.boot"
 
 const envNydusBuilder = "NYDUS_BUILDER"
 const envNydusWorkdir = "NYDUS_WORKDIR"
-
-type Layer struct {
-	// Digest represents the hash of whole tar blob.
-	Digest digest.Digest
-	// ReaderAt holds the reader of whole tar blob.
-	ReaderAt content.ReaderAt
-}
-
-type ConvertOption struct {
-	// RafsVersion specifies nydus format version, possible values:
-	// `5`, `6` (EROFS-compatible), default is `5`.
-	RafsVersion string
-	// ChunkDictPath holds the bootstrap path of chunk dict image.
-	ChunkDictPath string
-	// PrefetchPatterns holds file path pattern list want to prefetch.
-	PrefetchPatterns string
-}
-
-type MergeOption struct {
-	// ChunkDictPath holds the bootstrap path of chunk dict image.
-	ChunkDictPath string
-	// PrefetchPatterns holds file path pattern list want to prefetch.
-	PrefetchPatterns string
-	// WithTar puts bootstrap into a tar stream (no gzip).
-	WithTar bool
-}
 
 func getBuilder() string {
 	builderPath := os.Getenv(envNydusBuilder)
