+++ pkg/utils/mount/mount_other.go
@@ -1,14 +1,17 @@
+//go:build !linux
 // +build !linux
 
 package mount
+
+import "fmt"
 
 type Mounter struct {
 }
 
 func (m *Mounter) Umount(target string) error {
-	return nil
+	return fmt.Errorf("not implemented")
 }
 
 func (m *Mounter) IsLikelyNotMountPoint(file string) (bool, error) {
-	return true, nil
+	return false, fmt.Errorf("not implemented")
 }
