--- snapshot/snapshot_darwin.go
@@ -1,8 +1,0 @@
-// +build !linux
-
-package snapshot
-
-func getSupportsDType(dir string) (bool, error){
-	return true, nil
-}
-
